Added a StyleSheet class
Following as much as possible conventions familiar to html and MS Word
users.

The Template class uses the default StyleSheet, but that sheet can be
extended and used instead. Template class has meta-programming to build
methods to use in the pdf (ie. h1 "Some Text"), as well as get the
styles (h1_style)