Refined meta-programming and use of class variables
- StyleSheet
All methods that end in "_defaults" now have corresponding methods for
"_style" that are easily overridden in the template or other stylesheets
without having to merge in the related defaults. Makes it slightly
easier for developers of custom style sheets.

At the same time, and array of symbols of each style is set. This allows
the Template's method missing to see if it should create a method that
renders text in the selected style.

- Builder
The uses_template method now makes use of a class instance variable.
Makes the code much easier to understand, and simplifies the template
method. Rather than dynamically creating the method at run time, it
exists by default and raises and error if the template class has not
been set.

- Template
Similar to the uses_template method in the Builder, the same technique
is used for uses_style_sheet.

Changed method names for each component so there is no confusion with
the style text methods. ie. footer becomes footer_content, because the
footer_style creates a method called footer that writes text to the pdf
using the footer style.