StyleSheet is now a Module, not a Class
- The way it was being used by the Template lends itself much better as
a module that is included in the Template.

- Rather than meta programming the x_style methods from the defaults,
the defaults are no longer needed, so styles are defined directly. This
makes it easy to merge them to new styles in Templates.

- All styles are now StyleHash objects instead of vanilla Hash. This
adds extra keys for Prawn.formatted_text without forcing the user to
remember to define them (:font_size returns the :size value)